Fixed Hits Always Visible
I found it strange that the hit numbers were visible through walls and pretty much gave you an unintentional wallhack. 
I'm not sure if this is intentional but I've added a simple trace to check if LocalPlayer can actually see the Hit Entity.

I know it's probably not the best code and that it probably shouldn't be all in one line, but it's the best I can do at 4AM with a headache.